requirements.txt
google-genai
fastapi
uvicorn
openai==0.28
requests
redis
faiss-cpu 
numpy